Add card button started
Started work on the AddCard button. Need to get onClick working and then
should be complete. Also refactored some of the file structure to add in
the icon for AddCard. Refacted part of Card so the data is directly in
the component and not a global variable, but it is still example data.